更改模板信息

+++ devpl-backend/devpl-server/src/main/java/io/devpl/backend/domain/vo/TemplateSelectVO.java
@@ -13,7 +13,7 @@ public class TemplateSelectVO {
     /**
      * 模板ID
      */
-    private Integer templateId;
+    private Long templateId;
 
     /**
      * 模板名称
